#160 Refactor: change data, define hideError untils
